Move project cloud sync to $lib/project
